// src/app/model/FrecuenteReqModel.model.ts
export class FrecuenteReqModel{
 
    idfrecuente?: number;
    id_usuario?: number;
    id_empresa?: number;
    id_region?: number;
    id_provincia?: number;
    id_comuna?: number;
    direccion?: string;
    nombre?: string;
    email?: string;
    telefono?: string;
    rut?: number;
    dv?: string;
    fecha_creacion?: string;
    fecha_actualizacion?: string;

    constructor(){
          
    }
}